Fix ArrayIndexOutOfBoundsException.
When there is a trailing // comment on a invocation line the everything list only has the one comment here.  do a index check before a get so the ArrayIndexOutOfBoundsException isn't through when it shouldn't.  This would happen if the everything list only contains comments, not sure how this comes to be.