Enabled access to the tokens after parsing
Enabled the COMMON_TOKEN_ACTION=true option in the java_1_8.jj file.

Implemented a CommonTokenAction(Token token) method in the
TOKEN_MGR_DECLS that collects encountered tokens in a list.

Added a getter for this list of tokens in the ASTParser.

Made ASTParser public so that it may be used directly, not just using
the JavaParser wrapper class.

Made JavaParser.insertComments(...) public so that it may be used to
attach comments to a CompilationUnit that has been parsed using the
ASTParser directly.