JavaParser refactoring
The main parsing methods of the JavaParser now do not return a Node
subclass anymore, but instead an instance of the new ParseResult class.
This class wraps a Node plus some extra information about the parsing
process. At the moment the additional information is the list of tokens
that the parser has built while parsing the source code. It could
however be extended in the future to provide more information about the
parsing process.

Changes in detail:

Enabled the COMMON_TOKEN_ACTION=true option in the java_1_8.jj file.

Implemented a CommonTokenAction(Token token) method in the
TOKEN_MGR_DECLS that collects encountered tokens in a list.

Added a getter for this list of tokens in the ASTParser.

Added the ParseResult class as a wrapper for Node and List<Token>.

Changed parsing methods of JavaParser to return ParseResults instead of
Node subclasses.

Fixed testing source code to work with the new JavaParser API.

Cleaned up the JavaParser's parse() methods that parse a
CompilationUnit:
* Consistent overloaded versions for all input types (String,
  InputStream, File and Reader)
* Only the parse() method that accepts a String as input and mandatory
  'encoding' and 'considerComments' arguments contains the logic of how
  to actually parse and create a result. All other version just forward
  to this method by building a String and passing it to this method.
* Added JavaDoc for some undocumented public methods.
* Fixed JavaDoc: new return types.