JavaParser: Renamed methods with new return type
Renamed the methods that returned a ParseResult instead of a Node subclass
after the previous commit. The are now called 'parseToResult' and
'parse*ToResult'.

Recreated the former method signatures by forwarding to the
parseToResult methods and unpacking the Node from the ParseResult as
return value.

Fixed JavaDoc return type information.