New classes: InstanceJavaParser, CommentsInserter
Added the class 'InstanceJavaParser' as a thin wrapper around
ASTParser with a few convenience methods for parsing CompilationUnits,
Blocks, Statements, etc.

Added the class 'CommentsInsert' to encapsulate code for assigning
comments to nodes of the AST.

Refactored JavaParser to be based on both InstanceJavaParser and
CommentsInserter. This keeps the established API while having the logic
for making calls to the ASTParser in a single location.

Enabled the COMMON_TOKEN_ACTION=true option in the java_1_8.jj file.

Implemented a CommonTokenAction(Token token) method in the
TOKEN_MGR_DECLS that collects encountered tokens in a list.

Added a getter for this list of tokens in the ASTParser.