Fixed randomly failing tests.
The tests failed, because the surefire plugin executes up to 4 test methods in parallel (configured in pom.xml), and in these test classes all test methods modify the static configuration. Depending on the interleaving of the execution, this makes the test fail, because the type solver needed to proceed in one method has already been changed by another method.