Improve type resolution for duplicate names
When resolving types inside a ClassOrInterfaceDeclarationContext the
types found in extends and implements where searched before local
members. This caused problems for code such as:

import other.I;

class A implements I {
    interface I { }

    void foo(I i) { }
}

In the above the first parameter type for "foo" should be A.I and not
other.I.

Moving the code in ClassOrInterfaceDeclarationContext that searched
extends and implements first to JavaParserTypeDeclarationAdapter and
after searching local members showed the original problem of using
ClassOrInterfaceDeclarationContext when resolving extends and
implements as now A would instead implement A.I. Made a new fix for
that by introducing ClassOrInterfaceDeclarationExtendsContext which
only seaches in the parent context with one exception: type parameters.