Update PULL_REQUEST_TEMPLATE.md
Here is a short proposal for #3104 .

I don't want it to be overly rigid/bureaucratic or otherwise get in the way.

Instead, just act as a reminder for common things to look out for (e.g., ensuring that changes to one solver have been made in all solvers to keep them in sync where possible).

Comments/changes/objections/alternative proposals etc. are all appreciated!