Name Node string caching
The method `Name.asString()` is frequently used. However, calculating
its value can be quite expensive, especially when used with qualified
identifiers. In that case calling asString will always create a new
String object from the concatenation of the different parts of the
qualified name and the "." strings. This takes a considerable
amount of time and also puts pressure on the memory used.

On the other hand, recalculating the asString() value is not
necessary when we cache its value and re-use the value the
next time it is needed (and invalidate the cache when required.)

To enable this cache set the system property

- com.github.javaparser.performance.cache-node-name-string

to `true` before loading the JavaParser, or more specifically,
before loading the Name class.