Fast Name Node equals
When analyzing Java code using the JavaParser comparing
Name nodes using the equals method is a very frequent operation.
This comparison is done through the generic approach implemented in
JavaParser via the EqualsVisitor and HashCodeVisitor. In that process
the nullable field values of Name (identifier and qualifier) are wrapped
in an Optional before the actual comparison takes place. (*)

Creating multiple Optionals for every comparison between Name nodes
has a significant performance impact and also puts pressure on the
memory used (the Optionals will only have a very short live time but
will trigger frequent garbage collections, especially when working
on large code bases).

To avoid these extra Optionals the comparison between Name nodes
is now implemented by directly working on the field values, and checking
for null if necessary. This implementation in Name also speeds up the code
by avoiding overhead introduced by the Visitor pattern.

---
(*) A similar wrapping in Optional also occurs for the comment property,
inherited from Node. In this refactoring, however, the comment property
is not considered. First, this was less code is involved in the refactorying.
Secondly, it can be assumed that most Name nodes that are not equal
will already differ in the identifier and qualifier, so the comment check
is rarely executed.