Fix: #3725 JavaParserFacade var type in for-each loop cannot be resolved
Before this fix it was not possible to resolve the type of the variable
used in a for-each loop when the type of the variable was defined as `var`.
This was because the original code only took care of the `var with
initializer` case (e.g. `var i = "hello";`).

The fix now also covers the usage of a `var` in a for-each loop, e.g.
as in `for (var s: names) {...}`. The expression at the right side
of the `:` must either be an array or an Iterable.

The bug was originally reported in
https://github.com/javaparser/javaparser/issues/3725