Adapt NodeList.getFirst/getLast to Java 21
Java 21 reserves List<T> methods getFirst() and getLast() to return
objects with type T. The original methods added to return an Optional<T>
are removed in favour of using the default implementation.

Existing callsites are easy to convert: either check size() first if the
callsite was using ifPresent(), or directly get the object in case the
code was already calling getFirst().get() or getLast().get().